Revamp how we select which content to publish.
When first publishing content, instead of blindly replacing the first
piece of content that matches the notebook title, let the user choose
to either create a new piece of content or select content to overwrite
from the top 5 search results.

If the user does not change the title of the notebook then we take it
as a sign that the user wants to publish the notebook to the same
location.

If the user changes the title we go back to the search/choose-content
to overwrite flow.